lifesafe-flask/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
flask-jwt-extended==4.4.4
pymongo==4.5.0
passlib==1.7.4
python-dotenv==1.0.0
gunicorn==20.1.0
google-auth==2.22.0          # optional; include only if using Google sign-in verification